chore: rename editable-name to editable-name/name-editor
Refs: #279
