+++ src/components/pages/account-page/current-user-info/editable-name/name-editor/change-name.api.ts
@@ -20,7 +20,7 @@ export async function changeName({ csrfToken, ...bodyData }: Params) {
       },
       body: JSON.stringify(bodyData),
       credentials: 'include',
-    }
+    },
   )
   if (fetchDataResult instanceof Error) {
     return fetchDataResult
